Доработка адаптера и установка значений lastUpdate и imageUrl в CoinMapper.kt, CoinInfoAdapter.kt, CoinDetailActivity.kt.

+++ app/src/main/java/com/dscreate_app/crip/data/database/CoinInfoDbModel.kt
@@ -2,10 +2,6 @@ package com.dscreate_app.crip.data.database
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.dscreate_app.crip.data.network.ApiFactory.BASE_IMAGE_URL
-import com.dscreate_app.crip.presentation.utils.convertTime
-import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
 
 @Entity(tableName = "full_price_list")
 data class CoinInfoDbModel(
@@ -13,7 +9,7 @@ data class CoinInfoDbModel(
     val fromSymbol: String,
     val toSymbol: String,
     val price: String,
-    val lastUpdate: Int,
+    val lastUpdate: Long,
     val highDay: Double,
     val lowDay: Double,
     val lastMarket: String,
